Massive loader boot optimizations, also now more exception-free
Now uses std::initializer_list<HolyFunc> to register the functions and
doesn't use any vectors/strings for FFI, only calls NewVirtualChunk()
once and copies all the stuff at once so it uses less memory and it's faster.

Also now uses std::error_code overloads of <filesystem> routines so you
won't run into std::terminate()s if something goes fucky wucky with
vfs stuff(the only scenario where this falls apart is where
std::filesystem::remove_all fails to allocate memory which will most
likely not happen and also would cause other issues such as the PC
freezing before such an error in TINE even happens)

+++ src/sdl_window.hxx
@@ -2,22 +2,21 @@
 
 #include <string>
 
-#include <stddef.h>
-#include <stdint.h>
+#include "types.h"
 
 void        SetClipboard(char const *text);
 std::string ClipboardText();
 void        NewDrawWindow();
-void        DrawWindowUpdate(uint8_t *colors, uintptr_t internal_width);
+void        DrawWindowUpdate(u8 *colors, u64 internal_width);
 void        InputLoop(bool *off);
 
 extern "C" union bgr_48 {
-  uint64_t i;
+  u64 i;
   struct [[gnu::packed]] {
-    uint16_t b, g, r, pad;
+    u16 b, g, r, pad;
   };
 };
-void GrPaletteColorSet(uint64_t, bgr_48);
+void GrPaletteColorSet(u64, bgr_48);
 
 void SetKBCallback(void *fp, void *data);
 void SetMSCallback(void *fp);
